api/requirements.txt
fastapi
matplotlib
numpy
pillow
torch
torch-summary
torchview
torchvision
uvicorn
